api/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2.2
psycopg2-binary>=2.9
gunicorn>=20.1.0
python-dotenv>=1.0.0
drf-yasg>=1.20.0
